+++ src/store/pools/index.js
@@ -1,7 +1,7 @@
 import { reactive } from 'vue'
-
-import DAI_USDT from './DAI_USDT'
+// TODO: 
+// import DAI_USDT from './DAI_USDT'
 
 export default reactive({
-  DAI_USDT,
+  // DAI_USDT,
 })