+++ src/store/tokens/DAI_USDT.js
@@ -1,9 +1,0 @@
-import { ModelToken } from '../../models'
-import { getDotenvAddress } from '../helpers/methods'
-
-export default ModelToken.create({
-  code: 'DAI_USDT',
-  address: getDotenvAddress('DAI_USDT_TOKEN'),
-  isLpt: true,
-  isInfiniteAllowance: true
-})